Implement saving an event

+++ EventTracker/ViewModels/EventFormViewModel.cs
@@ -11,7 +11,7 @@ namespace EventTracker.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Category { get; set; }
+        public byte Category { get; set; }
         public IEnumerable<Category> Categories{ get; set; }
     }
 }
